(fix/feat) Fixed normal directions + growth ratio capability
Inverted the splines for upper and lower side, now the loop is clockwise and the normals are pointing in the same directions. Added the growth ratio capability which replaces the farfield mesh size with a growth ratio. This parameter allows to easily control the mesh size in the domain and is very useful for transonic computations in DART.